+++ app/src/main/res/layout/fragment_ticket_list.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".reservation.TicketListFragment">
+        tools:context=".ticket.TicketListFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rv_reservation"
